♻️ 코드 수정 : 패키지 구조 변경
