+++ healthtart/src/main/java/com/dev5ops/healthtart/equipment_per_gym/repository/EquipmentPerGymRepository.java
@@ -1,6 +1,6 @@
 package com.dev5ops.healthtart.equipment_per_gym.repository;
 
-import com.dev5ops.healthtart.equipment_per_gym.aggregate.EquipmentPerGym;
+import com.dev5ops.healthtart.equipment_per_gym.domain.entity.EquipmentPerGym;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
