--- healthtart/src/main/java/com/dev5ops/healthtart/exercise_equipment/aggregate/vo/request/RequestEditEquipmentVO.java
@@ -1,19 +1,0 @@
-package com.dev5ops.healthtart.exercise_equipment.aggregate.vo.request;
-
-import com.dev5ops.healthtart.equipment_per_gym.aggregate.EquipmentPerGym;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class RequestEditEquipmentVO {
-    private String exerciseEquipmentName;
-    private String bodyPart;
-    private String exerciseDescription;
-    private String exerciseImage;
-    private String recommendedVideo;
-}
