--- healthtart/src/main/java/com/dev5ops/healthtart/exercise_equipment/aggregate/vo/request/RequestRegisterEquipmentVO.java
@@ -1,16 +1,0 @@
-package com.dev5ops.healthtart.exercise_equipment.aggregate.vo.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class RequestRegisterEquipmentVO {
-    private String exerciseEquipmentName;
-    private String bodyPart;
-    private String exerciseDescription;
-    private String exerciseImage;
-    private String recommendedVideo;
-}
