+++ healthtart/src/main/java/com/dev5ops/healthtart/exercise_equipment/domain/vo/response/ResponseEditEquipmentVO.java
@@ -1,0 +1,19 @@
+package com.dev5ops.healthtart.exercise_equipment.domain.vo.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class ResponseEditEquipmentVO {
+    private String exerciseEquipmentName;
+    private String bodyPart;
+    private String exerciseDescription;
+    private String exerciseImage;
+    private String recommendedVideo;
+    private LocalDateTime updatedAt;
+}
