+++ healthtart/src/main/java/com/dev5ops/healthtart/exercise_equipment/domain/vo/response/ResponseFindEquipmentVO.java
@@ -1,0 +1,16 @@
+package com.dev5ops.healthtart.exercise_equipment.domain.vo.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class ResponseFindEquipmentVO {
+    private String exerciseEquipmentName;
+    private String bodyPart;
+    private String exerciseDescription;
+    private String exerciseImage;
+    private String recommendedVideo;
+}
