+++ healthtart/src/main/java/com/dev5ops/healthtart/exercise_equipment/repository/ExerciseEquipmentRepository.java
@@ -1,7 +1,6 @@
 package com.dev5ops.healthtart.exercise_equipment.repository;
 
-import com.dev5ops.healthtart.exercise_equipment.aggregate.ExerciseEquipment;
-import com.dev5ops.healthtart.gym.aggregate.Gym;
+import com.dev5ops.healthtart.exercise_equipment.domain.entity.ExerciseEquipment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
