--- healthtart/src/main/java/com/dev5ops/healthtart/gym/aggregate/vo/response/ResponseEditGymVO.java
@@ -1,17 +1,0 @@
-package com.dev5ops.healthtart.gym.aggregate.vo.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class ResponseEditGymVO {
-    private String gymName;
-    private String address;
-    private String businessNumber;
-    private LocalDateTime updatedAt;
-}
