+++ healthtart/src/main/java/com/dev5ops/healthtart/gym/domain/vo/request/RequestEditGymVO.java
@@ -1,0 +1,14 @@
+package com.dev5ops.healthtart.gym.domain.vo.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class RequestEditGymVO {
+    private String gymName;
+    private String address;
+    private String businessNumber;
+}
